+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Deprecation warning on invalid constructor in `HttpPartFormatter` class, by @HardNorth
+- Cookie utility methods in `HttpFormatUtils` class, by @HardNorth
 ### Changed
 - `jsoup` version update to fix vulnerabilities, by @HardNorth